Use DjangoApp hook
Since graphite is a Django app, we should be able to use the Django
hook.  We are supplying our own manage.py so that it loads the right
settings (as graphite uses an old one that doesn't).

Moved some templates to the right places so that the various hooks pick
them up automatically.